// frontend/src/types/json/service_sosmed_achievements.d.ts
export interface ServiceSosmedAchievements {
  id: number;
  documentId: string;
  achievement: string;
  parameter: string;
  background_color: string;
  text_color: string;
  createdAt: string;
  updatedAt: string;
  publishedAt: string;
  image: Image
}

export interface Image {
  id: number;
  documentId: string;
  name: string;
  alternativeText: string | null;
  caption: string | null;
  width: number;
  height: number;
  formats: unknown;
  hash: string;
  ext: string;
  mime: string;
  size: number;
  url: string;
  previewUrl: string | null;
  provider: string;
  provider_metadata: unknown | null;
  createdAt: string;
  updatedAt: string;
  publishedAt: string;
}